Eliminate use of isolates in parseCoverage.
Isolates were a useful performance booster when VM coverage was collected via
the `--coverage-dir` VM flag and potentially large numbers of coverage JSON
files needed to be parsed. With the elimination of that flag in 1.16, coverage
is always emitted to a single JSON file. This change also significantly
simplifies the code.