Test coverage for async function call (#199)
Adds a test to verify expected coverage is generated for an asynchronous
function with two branches, only one of which is executed.

Line 21 is currently marked coverable and not hit due to a VM bug. Once
https://github.com/dart-lang/sdk/issues/31222 is fixed and the new VM
released to stable, we should remove the test expectation for that line
and close https://github.com/dart-lang/coverage/issues/196.