Closing braces on async functions are not coverable (#201)
This eliminates test coverage expectations for the closing brace of
async functions on Dart VMs versioned 2.0 and higher. Due to a bug, the
closing brace of async function blocks was emitted as coverable by the
VM service (it shimmed in completeError as a hidden token on the closing
brace). See: https://github.com/dart-lang/sdk/issues/31222

As of https://dart-review.googlesource.com/c/sdk/+/16605, that issue is
resolved in the VM.