Update tests for Dart SDK 2.0.0-dev.64.1
The --checked flag has been removed in Dart SDK 2.0.0-dev.64.1. This
changes the default for the `checked` parameter in runAndCollect() to
false. In Dart 2, checked mode is no longer necessary, since type checks
are done statically at compile time. Support for the --checked flag will
be removed in a future version of package:coverage.