Add version of parseJson that can cache file reads and parses
When checkIgnoredLines is true parseJson reads files to find
ignore data for that file.
If calling parseJson several times for the same project,
as for instance done in flutter, this adds significant overhead.

This commit adds a new function - parseJsonSync - which is passed
a cache of the lookups. This way future lookups for the same
file will be faster. The new function is also sync avoiding any
potential race conditions when filling up the cache.